http://cvmfs.readthedocs.io/en/latest/cpt-replica.html
In order to prevent accidental synchronization from a repository, the
Stratum 0 repository maintainer has to create a .cvmfs_master_replica
file in the HTTP root directory. This file is created by default when a
new repository is created. Note that replication can thrash caches that
might exist between Stratum 1 and Stratum 0. A direct connection is
therefore preferable.